requirements.txt
asgiref==1.1.2
attrs==17.2.0
autobahn==17.9.2
Automat==0.6.0
channels==1.1.8
constantly==15.1.0
daphne==1.3.0
Django==1.11.5
django-guardian==1.4.9
django-polymorphic==1.3
djangorestframework==3.7.7
hyperlink==17.3.1
incremental==17.5.0
pytz==2017.2
six==1.11.0
Twisted==17.5.0
txaio==2.8.2
websocket-client==0.44.0
websockets==4.0.1
zope.interface==4.4.2
asgiref==1.1.2
attrs==17.2.0
autobahn==17.9.2
Automat==0.6.0
channels==1.1.8
constantly==15.1.0
coverage==4.4.2
daphne==1.3.0
Django==1.11.5
django-cors-headers==2.1.0
django-guardian==1.4.9
django-polymorphic==1.3
django-webpack-loader==0.5.0
djangorestframework==3.7.7
hyperlink==17.3.1
incremental==17.5.0
pkg-resources==0.0.0
pytz==2017.2
six==1.11.0
Twisted==17.5.0
txaio==2.8.2
websocket-client==0.44.0
websockets==4.0.1
zope.interface==4.4.2
